start implementing hash-based testing
The idea is to avoid redundant testing. This is achieved by storing
the outcome in a dictionary, with the hash of the tested file as the
key, and the result as the value. When the result is requested again
later, it can be retrieved from the dictionary, thus avoiding calculating
it again.